IMPLEMENT: Add changes for datatype conversion

+++ src/main/java/com/example/demo/models/dto/UserAdvanceSalaryDTO.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserAdvanceSalaryDTO {
-    private int userId;
-    private double advanceSalaryAmount;
+    private long userId;
+    private BigDecimal advanceSalaryAmount;
 }
